+++ data-model-api/src/main/java/com/provectus/prodobro/shared/avatar/Avatar.java
@@ -3,11 +3,11 @@ package com.provectus.prodobro.shared.avatar;
 
 public interface Avatar {
 
-    int getId();
+    Long getId();
 
     byte[] getBytea();
 
-    void setId(int id);
+    void setId(Long id);
 
     void setBytea(byte[] bytea);
 }
